Add DeliveryThreadManager to respawn thread
This is necessary for preforking servers such as Unicorn, since forking
causes the thread to be killed.  This case can be reproduced by initializing
a logglier instance in a Rails config and running the app under Unicorn with
`preload_app true`.  In that case, the thread would be killed, and the @queue
would become a blackhole where messages silently disappear.